Refactor cordova-serve to use Express.
This simplifies a lot of code, and also provides a more standardized and modular way to customize the server (via Express middleware).

Also changes cordova-serve to support multiple instances (i.e. multiple servers).

Note that this is a breaking change, so for next release we will need to update the version to 0.2.0 (and I have updates to cordova-browser and cordova-lib ('cordova serve' command) ready to go to make use of this change.